@@ -14,9 +14,12 @@
     ]
   },
   "require": {
-    "php": ">= 5.3.0",
-    "ext-curl": "*",
+    "php": ">= 8.1",
     "ext-libxml": "*",
-    "ext-simplexml": "*"
+    "ext-simplexml": "*",
+    "symfony/http-client": "^6.0"
+  },
+  "require-dev": {
+    "rector/rector": "^0.15.25"
   }
 }
