@@ -31,106 +31,108 @@
 define('DEBUG', true);
 define('PS_SHOP_PATH', 'http://www.myshop.com/');
 define('PS_WS_AUTH_KEY', 'ZQ88PRJX5VWQHCWE4EE7SQ7HPNX00RAJ');
-require_once('../PSWebServiceLibrary.php');
+require_once(__DIR__ . '/../PSWebServiceLibrary.php');
 
 // Here we use the WebService to get the schema of "customers" resource
-try
-{
+try {
     $webService = new PrestaShopWebservice(PS_SHOP_PATH, PS_WS_AUTH_KEY, DEBUG);
-    $opt = array('resource' => 'customers');
-    if (isset($_GET['Create']))
-        $xml = $webService->get(array('url' => PS_SHOP_PATH.'/api/customers?schema=blank'));
-    else
+    $opt = ['resource' => 'customers'];
+    if (isset($_GET['Create'])) {
+        $xml = $webService->get(['url' => PS_SHOP_PATH.'/api/customers?schema=blank']);
+    } else {
         $xml = $webService->get($opt);
+    }
+
     $resources = $xml->children()->children();
 
-} catch (PrestaShopWebserviceNotFoundException $exception) {
+} catch (PrestaShopWebserviceNotFoundException $prestaShopWebserviceNotFoundException) {
     echo 'Bad ID';
     exit;
-} catch (PrestaShopWebserviceUnauthorizedException $exception) {
+} catch (PrestaShopWebserviceUnauthorizedException) {
     echo 'Bad auth key';
     exit;
-} catch (PrestaShopWebserviceForbiddenException $exception) {
+} catch (PrestaShopWebserviceForbiddenException) {
     echo 'Not logged in';
     exit;
-} catch (PrestaShopWebserviceException $exception) {
-    echo 'Other error<br />'.$exception->getMessage();
+} catch (PrestaShopWebserviceException $prestaShopWebserviceException) {
+    echo 'Other error<br />'.$prestaShopWebserviceException->getMessage();
     exit;
 }
 
-if (count($_POST) > 0)
-{
-// Here we have XML before update, lets update XML
-    foreach ($resources as $nodeKey => $node)
-    {
+if ($_POST !== []) {
+    // Here we have XML before update, lets update XML
+    foreach ($resources as $nodeKey => $node) {
         $resources->$nodeKey = $_POST[$nodeKey];
     }
-    try
-    {
-        $opt = array('resource' => 'customers');
-        if ($_GET['Create'] == 'Creating')
-        {
+
+    try {
+        $opt = ['resource' => 'customers'];
+        if ($_GET['Create'] == 'Creating') {
             $opt['postXml'] = $xml->asXML();
             $xml = $webService->add($opt);
             echo "Successfully added.";
         }
-    } catch (PrestaShopWebserviceNotFoundException $exception) {
+    } catch (PrestaShopWebserviceNotFoundException $prestaShopWebserviceNotFoundException) {
         echo 'Bad ID';
-    } catch (PrestaShopWebserviceUnauthorizedException $exception) {
+    } catch (PrestaShopWebserviceUnauthorizedException) {
         echo 'Bad auth key';
-    } catch (PrestaShopWebserviceForbiddenException $exception) {
+    } catch (PrestaShopWebserviceForbiddenException) {
         echo 'Not logged in';
-    } catch (PrestaShopWebserviceException $exception) {
-        echo 'Other error<br />'.$exception->getMessage();
+    } catch (PrestaShopWebserviceException $prestaShopWebserviceNotFoundException) {
+        echo 'Other error<br />'.$prestaShopWebserviceNotFoundException->getMessage();
     }
 }
 
 // We set the Title
-echo '<h1>Customer\'s ';
-if (isset($_GET['Create'])) echo 'Creation';
-else echo 'List';
+echo "<h1>Customer's ";
+if (isset($_GET['Create'])) {
+    echo 'Creation';
+} else {
+    echo 'List';
+}
+
 echo '</h1>';
 
 // We set a link to go back to list if we are in creation
-if (isset($_GET['Create']))
+if (isset($_GET['Create'])) {
     echo '<a href="?">Return to the list</a>';
+}
 
-if (!isset($_GET['Create']))
+if (!isset($_GET['Create'])) {
     echo '<input type="button" onClick="document.location.href=\'?Create\'" value="Create">';
-else
+} else {
     echo '<form method="POST" action="?Create=Creating">';
+}
 
 echo '<table border="5">';
-if (isset($resources))
-{
+if (isset($resources)) {
 
-echo '<tr>';
-if (count($_GET) == 0)
-{
-    echo '<th>Id</th></tr>';
+    echo '<tr>';
+    if (count($_GET) == 0) {
+        echo '<th>Id</th></tr>';
 
-    foreach ($resources as $resource)
-    {
-        echo '<tr><td>'.$resource->attributes().'</td></tr>';
+        foreach ($resources as $resource) {
+            echo '<tr><td>'.$resource->attributes().'</td></tr>';
+        }
+    } else {
+        echo '</tr>';
+        foreach ($resources as $key => $resource) {
+            echo '<tr><th>'.$key.'</th><td>';
+            if (isset($_GET['Create'])) {
+                echo '<input type="text" name="'.$key.'" value=""/>';
+            }
+
+            echo '</td></tr>';
+        }
     }
-}
-else
-{
-    echo '</tr>';
-    foreach ($resources as $key => $resource)
-    {
-        echo '<tr><th>'.$key.'</th><td>';
-        if (isset($_GET['Create']))
-            echo '<input type="text" name="'.$key.'" value=""/>';
-        echo '</td></tr>';
-    }
-}
 
 }
+
 echo '</table><br/>';
 
-if (isset($_GET['Create']))
+if (isset($_GET['Create'])) {
     echo '<input type="submit" value="Create"></form>';
+}
 
 
 ?>
