@@ -31,109 +31,111 @@
 define('DEBUG', true);
 define('PS_SHOP_PATH', 'http://www.myshop.com/');
 define('PS_WS_AUTH_KEY', 'VIVELADIVISIONMOBILEDEPRESTASHOP');
-require_once('../PSWebServiceLibrary.php');
+require_once(__DIR__ . '/../PSWebServiceLibrary.php');
 
 // First : We always get the customer's list or a specific one
-try
-{
+try {
     $webService = new PrestaShopWebservice(PS_SHOP_PATH, PS_WS_AUTH_KEY, DEBUG);
-    $opt = array('resource' => 'customers');
-    if (isset($_GET['id']))
+    $opt = ['resource' => 'customers'];
+    if (isset($_GET['id'])) {
         $opt['id'] = $_GET['id'];
+    }
+
     $xml = $webService->get($opt);
 
     // Here we get the elements from children of customer markup which is children of prestashop root markup
     $resources = $xml->children()->children();
 
-} catch (PrestaShopWebserviceNotFoundException $exception) {
+} catch (PrestaShopWebserviceNotFoundException $prestaShopWebserviceNotFoundException) {
     echo 'Bad ID';
     exit;
-} catch (PrestaShopWebserviceUnauthorizedException $exception) {
+} catch (PrestaShopWebserviceUnauthorizedException) {
     echo 'Bad auth key';
     exit;
-} catch (PrestaShopWebserviceForbiddenException $exception) {
+} catch (PrestaShopWebserviceForbiddenException) {
     echo 'Not logged in';
     exit;
-} catch (PrestaShopWebserviceException $exception) {
-    echo 'Other error<br />'.$exception->getMessage();
+} catch (PrestaShopWebserviceException $prestaShopWebserviceException) {
+    echo 'Other error<br />'.$prestaShopWebserviceException->getMessage();
     exit;
 }
 
 // Second : We update the data and send it to the web service
-if (isset($_GET['id']) && isset($_POST['id'])) // Here we check id cause in every resource there's an id
-{
+if (isset($_GET['id']) && isset($_POST['id'])) { // Here we check id cause in every resource there's an id
     // Here we have XML before update, lets update XML with new values
-    foreach ($resources as $nodeKey => $node)
-    {
+    foreach ($resources as $nodeKey => $node) {
         $resources->$nodeKey = $_POST[$nodeKey];
     }
+
     // And call the web service
-    try
-    {
-        $opt = array('resource' => 'customers');
+    try {
+        $opt = ['resource' => 'customers'];
         $opt['putXml'] = $xml->asXML();
         $opt['id'] = $_GET['id'];
         $xml = $webService->edit($opt);
         // if WebService don't throw an exception the action worked well and we don't show the following message
         echo "Successfully updated.";
-    } catch (PrestaShopWebserviceNotFoundException $exception) {
+    } catch (PrestaShopWebserviceNotFoundException $prestaShopWebserviceNotFoundException) {
         echo 'Bad ID';
-    } catch (PrestaShopWebserviceUnauthorizedException $exception) {
+    } catch (PrestaShopWebserviceUnauthorizedException) {
         echo 'Bad auth key';
-    } catch (PrestaShopWebserviceForbiddenException $exception) {
+    } catch (PrestaShopWebserviceForbiddenException) {
         echo 'Not logged in';
-    } catch (PrestaShopWebserviceException $exception) {
-        echo 'Other error<br />'.$exception->getMessage();
+    } catch (PrestaShopWebserviceException $prestaShopWebserviceNotFoundException) {
+        echo 'Other error<br />'.$prestaShopWebserviceNotFoundException->getMessage();
     }
 }
 
 // UI
 
 // We set the Title
-echo '<h1>Customer\'s ';
-if (isset($_GET['id'])) echo 'Update';
-else echo 'List';
+echo "<h1>Customer's ";
+if (isset($_GET['id'])) {
+    echo 'Update';
+} else {
+    echo 'List';
+}
+
 echo '</h1>';
 
 // We set a link to go back to list if we are in customer's details
-if (isset($_GET['id']))
+if (isset($_GET['id'])) {
     echo '<a href="?">Return to the list</a>';
+}
 
-if (isset($_GET['id']))
+if (isset($_GET['id'])) {
     echo '<form method="POST" action="?id='.$_GET['id'].'">';
+}
+
 echo '<table border="5">';
-if (isset($resources))
-{
+if (isset($resources)) {
 
-echo '<tr>';
-if (!isset($_GET['id']))
-{
-    //Show list of customers
-    echo '<th>Id</th><th>More</th></tr>';
-    foreach ($resources as $resource)
-    {
-        echo '<td>'.$resource->attributes().'</td><td>'.
-        '<a href="?id='.$resource->attributes().'">Update</a>&nbsp;'.
-        '</td></tr>';
+    echo '<tr>';
+    if (!isset($_GET['id'])) {
+        //Show list of customers
+        echo '<th>Id</th><th>More</th></tr>';
+        foreach ($resources as $resource) {
+            echo '<td>'.$resource->attributes().'</td><td>'.
+            '<a href="?id='.$resource->attributes().'">Update</a>&nbsp;'.
+            '</td></tr>';
+        }
+    } else {
+        //Show customer form
+        echo '</tr>';
+        foreach ($resources as $key => $resource) {
+            echo '<tr><th>'.$key.'</th><td>';
+            echo '<input type="text" name="'.$key.'" value="'.$resource.'"/>';
+            echo '</td></tr>';
+        }
     }
-}
-else
-{
-    //Show customer form
-    echo '</tr>';
-    foreach ($resources as $key => $resource)
-    {
-        echo '<tr><th>'.$key.'</th><td>';
-        echo '<input type="text" name="'.$key.'" value="'.$resource.'"/>';
-        echo '</td></tr>';
-    }
-}
 
 }
+
 echo '</table><br/>';
 
-if (isset($_GET['id']))
+if (isset($_GET['id'])) {
     echo '<input type="submit" value="Update"></form>';
+}
 
 
 ?>
