@@ -31,64 +31,61 @@
 define('DEBUG', true);                                            // Debug mode
 define('PS_SHOP_PATH', 'http://www.myshop.com/');                            // Root path of your PrestaShop store
 define('PS_WS_AUTH_KEY', 'ZQ88PRJX5VWQHCWE4EE7SQ7HPNX00RAJ');    // Auth key (Get it in your Back Office)
-require_once('../PSWebServiceLibrary.php');
+require_once(__DIR__ . '/../PSWebServiceLibrary.php');
 
 // Here we make the WebService Call
-try
-{
+try {
     $webService = new PrestaShopWebservice(PS_SHOP_PATH, PS_WS_AUTH_KEY, DEBUG);
     // Here we set the option array for the Webservice : we want customers resources
     $opt['resource'] = 'customers';
     // We set an id if we want to retrieve infos from a customer
-    if (isset($_GET['id']))
-        $opt['id'] = (int)$_GET['id']; // cast string => int for security measures
+    if (isset($_GET['id'])) {
+        $opt['id'] = (int)$_GET['id'];
+    } // cast string => int for security measures
 
     // Call
     $xml = $webService->get($opt);
 
     // Here we get the elements from children of customer markup which is children of prestashop root markup
     $resources = $xml->children()->children();
-} catch (PrestaShopWebserviceNotFoundException $exception) {
+} catch (PrestaShopWebserviceNotFoundException) {
     echo 'Bad ID';
-} catch (PrestaShopWebserviceUnauthorizedException $exception) {
+} catch (PrestaShopWebserviceUnauthorizedException) {
     echo 'Bad auth key';
-} catch (PrestaShopWebserviceForbiddenException $exception) {
+} catch (PrestaShopWebserviceForbiddenException) {
     echo 'Not logged in';
-} catch (PrestaShopWebserviceException $exception) {
-    echo 'Other error<br />'.$exception->getMessage();
+} catch (PrestaShopWebserviceException $prestaShopWebserviceException) {
+    echo 'Other error<br />'.$prestaShopWebserviceException->getMessage();
 }
 
 // We set the Title
 echo '<h1>Customers ';
-if (isset($_GET['id']))
+if (isset($_GET['id'])) {
     echo 'Details';
-else
+} else {
     echo 'List';
+}
+
 echo '</h1>';
 
 // We set a link to go back to list if we are in customer's details
-if (isset($_GET['id']))
+if (isset($_GET['id'])) {
     echo '<a href="?">Return to the list</a>';
+}
 
 echo '<table border="5">';
 // if $resources is set we can lists element in it otherwise do nothing cause there's an error
-if (isset($resources))
-{
-    if (!isset($_GET['id']))
-    {
+if (isset($resources)) {
+    if (!isset($_GET['id'])) {
         echo '<tr><th>Id</th><th>More</th></tr>';
-        foreach ($resources as $resource)
-        {
+        foreach ($resources as $resource) {
             // Iterates on the found IDs
             echo '<tr><td>'.$resource->attributes().'</td><td>'.
             '<a href="?id='.$resource->attributes().'">Retrieve</a>'.
             '</td></tr>';
         }
-    }
-    else
-    {
-        foreach ($resources as $key => $resource)
-        {
+    } else {
+        foreach ($resources as $key => $resource) {
             // Iterates on customer's properties
             echo '<tr>';
             echo '<th>'.$key.'</th><td>'.$resource.'</td>';
@@ -96,6 +93,7 @@ if (isset($resources))
         }
     }
 }
+
 echo '</table>';
 ?>
 </body></html>
